Reorganize Codebase
- [+] refactor: move keyidentifier package under authentication/crypto directory
